+++ tuneapi/utils/env.py
@@ -8,8 +8,8 @@ class _ENV:
     def __init__(self):
         self.vars_called = set()
 
-    TUNE_API_KEY = lambda x="": os.getenv("TUNE_API_KEY", x)
-    TUNE_ORG_ID = lambda x="": os.getenv("TUNE_ORG_ID", x)
+    TUNEAPI_TOKEN = lambda x="": os.getenv("TUNEAPI_TOKEN", x)
+    TUNEORG_ID = lambda x="": os.getenv("TUNEORG_ID", x)
 
     def __getattr__(self, __name: str):
         self.vars_called.add(__name)
@@ -25,6 +25,6 @@ Usage:
 .. code-block:: python
 
     from tuneapi.utils.env import ENV
-    TUNE_API_KEY = ENV.TUNE_API_KEY()   # get defined values
+    TUNEAPI_TOKEN = ENV.TUNEAPI_TOKEN()   # get defined values
     MY_ENV = ENV.MY_ENV()               # get your arbitrary env var
 """